Allow keeping old rpm log in package directory
With this change, it is now possible to store the old rpm log inside the
package directory instead of using a separate directory (usually /misc/
directory). This can be done by changing the oldurl setting in
mgarepo.conf to '.' or anything starting with './'.

It becomes possible to have different import logs if using different
packages in different versions of the distribution instead of one log
file shared for all versions of the distribution.

git-svn-id: svn+ssh://svn.mageia.org/svn/soft/build_system/mgarepo/trunk@6559 5b7c3c08-40e5-403b-9995-ace06908e4af